+++ src/group/datastore/index.js
@@ -1,4 +1,4 @@
-import currentGroup from '@/group/datastore/currentGroup'
+import currentGroup, { plugin as currentGroupPlugin } from '@/group/datastore/currentGroup'
 import timezones from '@/group/datastore/timezones'
 
 export default {
@@ -6,4 +6,7 @@ export default {
     currentGroup,
     timezones,
   },
+  plugins: [
+    currentGroupPlugin,
+  ],
 }