+++ src/sidenav/components/SidenavTitle.vue
@@ -1,17 +1,25 @@
 <template>
   <q-toolbar v-if="$q.platform.is.mobile">
-    <q-toolbar-title v-if="currentGroup">{{ currentGroup.name }}</q-toolbar-title>
-    <CommunityFeed />
-    <LatestMessageButton @click="$emit('click')" />
-    <NotificationButton @click="$emit('click')" />
+    <q-toolbar-title>{{ currentGroup && currentGroup.name }}</q-toolbar-title>
+    <CommunityFeed
+      @click.native="$emit('click')"
+    />
+    <LatestMessageButton
+      v-if="isLoggedIn"
+      @click="$emit('click')"
+    />
+    <NotificationButton
+      v-if="isLoggedIn"
+      @click="$emit('click')"
+    />
     <LocaleSelect />
     <q-btn
-      v-if="currentGroup"
       flat
       dense
       round
       :to="{ name: 'groupsGallery' }"
       :title="$t('TOPBAR.CHANGE_GROUP')"
+      @click.native="$emit('click')"
     >
       <q-icon name="fas fa-exchange-alt" />
       <q-tooltip v-t="'TOPBAR.CHANGE_GROUP'" />
@@ -25,7 +33,13 @@ import CommunityFeed from '@/communityFeed/components/CommunityFeed'
 import LatestMessageButton from '@/messages/components/LatestMessageButton'
 import NotificationButton from '@/notifications/components/NotificationButton'
 
-import { QToolbar, QToolbarTitle, QBtn, QIcon, QTooltip } from 'quasar'
+import {
+  QToolbar,
+  QToolbarTitle,
+  QBtn,
+  QIcon,
+  QTooltip,
+} from 'quasar'
 
 import { mapGetters } from 'vuex'
 
@@ -44,6 +58,7 @@ export default {
   computed: {
     ...mapGetters({
       currentGroup: 'currentGroup/value',
+      isLoggedIn: 'auth/isLoggedIn',
     }),
   },
 }