+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Please document your changes in this format:
 ### Fixed
 
 - Notification items would sometimes cause errors when related pickups haven't been loaded @tiltec
+- Sidenav title on mobile shows notification and message icons when logged out @tiltec
 
 ## [6.2.3] - 2018-10-26